Split commands with subcommands into separate Cannoli CLI classes.
